Revert "fix: Navigating to page, navigate to last used segment instead of can…"
This reverts commit b68f349eb6258e1f5089850a92b8ad13d90dfe5a.
