+++ app/client/cypress/e2e/Regression/Apps/EchoApiCMS_spec.js
@@ -75,7 +75,7 @@ describe(
 
     it("4. Send mail and verify post request body", function () {
       // navigating to canvas
-      PageLeftPane.switchSegment(PagePaneSegment.UI);
+      cy.xpath(appPage.pagebutton).click();
       cy.get(appPage.submitButton).should("be.visible");
       cy.xpath("//span[text()='3']").click({ force: true });
       cy.get(appPage.mailButton).closest("div").click();
@@ -89,7 +89,7 @@ describe(
         .type("Task completed", { force: true });
       cy.get(appPage.confirmButton).closest("div").click({ force: true });
       cy.get(appPage.closeButton).closest("div").click({ force: true });
-      PageLeftPane.switchSegment(PagePaneSegment.UI);
+      cy.xpath(appPage.pagebutton).click({ force: true });
       PageLeftPane.switchSegment(PagePaneSegment.Queries);
       cy.xpath(appPage.postApi).click({ force: true });
       cy.ResponseCheck("Test");
@@ -99,14 +99,14 @@ describe(
 
     it("5. Delete proposal and verify delete request body", function () {
       // navigating back to canvas
-      PageLeftPane.switchSegment(PagePaneSegment.UI);
+      cy.xpath(appPage.pagebutton).click({ force: true });
       cy.get(appPage.submitButton).closest("div").should("be.visible");
       cy.xpath("//span[text()='Dan.Wyman@hotmail.com']").click({ force: true });
       // deleting the proposal and asserting delete call's response
       cy.xpath(appPage.deleteButton).click({ force: true });
       cy.xpath(appPage.deleteTaskText).should("be.visible");
       cy.get(appPage.confirmButton).closest("div").click({ force: true });
-      PageLeftPane.switchSegment(PagePaneSegment.UI);
+      cy.xpath(appPage.pagebutton).click({ force: true });
       EditorNavigation.SelectEntityByName("delete_proposal", EntityType.Api);
       cy.ResponseCheck("Dan.Wyman@hotmail.com");
       cy.ResponseCheck("Recusan");
