+++ app/client/cypress/e2e/Regression/Apps/PromisesApp_spec.js
@@ -7,8 +7,6 @@ import {
 } from "../../../support/Objects/ObjectsCore";
 import EditorNavigation, {
   EntityType,
-  PageLeftPane,
-  PagePaneSegment,
 } from "../../../support/Pages/EditorNavigation";
 const commonlocators = require("../../../locators/commonlocators.json");
 
@@ -61,7 +59,7 @@ describe(
           shouldCreateNewJSObj: true,
         },
       );
-      PageLeftPane.switchSegment(PagePaneSegment.UI);
+      EditorNavigation.SelectEntityByName("Page1", EntityType.Page);
       cy.wait("@getConsolidatedData");
       // verify text in the text widget
 
@@ -115,7 +113,7 @@ describe(
       EditorNavigation.SelectEntityByName("JSObject1", EntityType.JSObject);
       jsEditor.SelectFunctionDropdown("clearStore");
       jsEditor.RunJSObj();
-      PageLeftPane.switchSegment(PagePaneSegment.UI);
+      EditorNavigation.SelectEntityByName("Page1", EntityType.Page);
       cy.xpath("//span[text()='Clear store']").click({ force: true });
       cy.get(".t--draggable-textwidget span")
         .eq(5)
