+++ app/client/cypress/e2e/Regression/ClientSide/ExplorerTests/Hide_Page_spec.js
@@ -1,6 +1,5 @@
 import EditorNavigation, {
   EntityType,
-  PageLeftPane,
 } from "../../../../support/Pages/EditorNavigation";
 
 import * as _ from "../../../../support/Objects/ObjectsCore";
@@ -17,7 +16,6 @@ describe(
         entityNameinLeftSidebar: "Page2",
         action: "Hide",
       });
-      PageLeftPane.switchToAddNew();
       cy.ClearSearch();
       _.deployMode.DeployApp();
       cy.get(".t--page-switch-tab").should("have.length", 2);
